Break the project into crates
This creates a workspace in the main cargo toml so that
we can break the entire project into 2 crates. One for the binary
and one for the library. This will solve a couple problems:
1. Allow the usage of doc tests in the library
2. Force a better public API for the binary to consume

It also separates the option parsing from the main app to create a
more logical structure between domain.